requirements.txt
urllib2
bs4
requests
gzip
os
json
hurry.filesize
io
